+++ run_on_coal/Utils/Pool.cpp
@@ -2,7 +2,7 @@
 
 #include "Utils/Pool.h"
 
-ROC::Pool::Pool(unsigned int f_size)
+ROC::Pool::Pool(size_t f_size)
 {
     m_size = f_size;
     m_poolData.assign(m_size, 0U);
@@ -16,7 +16,7 @@ ROC::Pool::~Pool()
 int ROC::Pool::Allocate()
 {
     int l_allocated = -1;
-    for(unsigned int i = m_minimal; i < m_size; i++)
+    for(size_t i = m_minimal; i < m_size; i++)
     {
         if(m_poolData[i] == 0U)
         {
@@ -27,7 +27,7 @@ int ROC::Pool::Allocate()
     }
     if(l_allocated == -1)
     {
-        for(unsigned int i = 0; i < m_minimal; i++)
+        for(size_t i = 0; i < m_minimal; i++)
         {
             if(m_poolData[i] == 0U)
             {
@@ -39,7 +39,7 @@ int ROC::Pool::Allocate()
     }
     if(l_allocated != -1)
     {
-        for(unsigned int i = static_cast<unsigned int>(l_allocated + 1); i < m_size; i++)
+        for(size_t i = static_cast<size_t>(l_allocated + 1); i < m_size; i++)
         {
             if(m_poolData[i] == 0U)
             {
@@ -50,7 +50,7 @@ int ROC::Pool::Allocate()
     }
     return l_allocated;
 }
-void ROC::Pool::Reset(unsigned int f_id)
+void ROC::Pool::Reset(size_t f_id)
 {
     if(f_id < m_size)
     {
